requirements.txt
flask
pycryptodome
pymysql
bcrypt
requests
multiprocessing
python-dotenv